fix: move ESI types to shared ports surface and fix formatting
Moves the canonical ESI type definitions from src/types/ethercat/ to
src/middleware/shared/ports/esi-types.ts so ports can import them
without violating the architecture layer rule (ports -> [utils, ports]).

The original file at src/types/ethercat/esi-types.ts becomes a
re-export barrel so the 30+ existing consumers continue working
without import changes.

Also fixes Prettier formatting on 6 EtherCAT files flagged by CI.
